Update Change.py
Added if statements in the clip_newDEM and clip_oldDEM functions inorder to make the shapefile entry by the user optional. If no shapefile is entered in the combo box the fucntions return the original DEM inputs as the fucntions values instead of running the clip rasker by mask layer native algorthim.